Exercise 7
We use the integers a, b, and n to create the following series:

(a + 2^0 * b), (a + 2^0 * b + 2^1 * b), (a + 2^0 * b + 2^1 * b ... + 2^n-1 * b)

You are given 1 queries in the form of a, b, and n. For each query, print the series corresponding to the given a, b, and n values as a single line of n space-separated integers.