Exercise 8
For each input variable n and appropriate primitive DataType, you must determine if the given primitives are capable of storing it. If yes, then print:

n can be fitted in:
* dataType

If there is more than one appropriate data type, print each one on its own line and order them by size.

If the number cannot be stored in one of the four aforementioned primitives, print the line:

n can't be fitted anywhere.