Exercise 23
Write the following code in the unlocked portion of your editor:

Create an array, A, capable of holding N integers.
Modify the code in the loop so that it saves each sequential value to its corresponding location in the array. For example, the first value must be stored in A0, the second value must be stored in An-1, and so on.