Exercise 25
You will be given two integers X and Y as input, you have to compute X/Y. If X and Y are not 32 bit signed integers or if Y is zero, exception will occur and you have to report it. Read sample Input/Output to know what to report in case of exceptions.