Exercise 36
You are required to compute the power of a number by implementing a calculator. Create a class MyCalculator which consists of a single method long power(int, int). This method takes two integers, N and P, as parameters and finds N^P. If either N or P is negative, then the method must throw an exception which says "N or P should not be negative". Also, if both N and P are zero, then the method must throw an exception which says "N and P should not be zero".